backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
huggingface-hub==0.19.4
reportlab==4.0.9 